requirements.txt
PyQt5>=5.15.0
pyotp>=2.6.0
qrcode>=7.3.1
cryptography>=3.4.7
Pillow>=8.2.0
pyperclip>=1.8.2
